requirements.txt
aws-cdk-lib>=2.213.0
constructs>=10.0.0,<11.0.0 # CDK v2 uses constructs 10.x
boto3>=1.40.18
